Redesign host models
- Split the host data into three KV entries. This fixes the concurrency
bug where if we update the IPv4 and the IPv6 with two consecutive
requests one of the address would get overridden by the second request
because of the propagation delay.
- Limit the number of queries with the Cloudflare API. We only send a
query when there is a change to be made.
- Limit the number of KV writes. The `last_update` and
`last_update_token_id` have disappeared because this was causing to many
writes and Cloudflare Workers free plan only allow for 1k writes greatly
limiting the number of requests we could make.
- The host object response has changed breaking compatibility with
previous version.